removes code duplication
